Adhere to the documentation, breaking FizzBuzz
Until now, the message bag used was kind of lenient — it would shuffle,
but only every so often. This allowed for clever programs to be executed
in an expected order. While this was useful, it goes against the spec
and spirit of Basil. The message bag has therefore be changed to always
shuffle itself upon popping.

It is unclear whether it is still possible to do things in order
somehow.